# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "-.",
        "title": "Go Whatsapp Notifier API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Kelana Chandra Helyandika",
            "url": "https://kelanach.xyz"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:3004",
    "basePath": "/api",
    "paths": {
        "/wa/logout": {
            "post": {
                "description": "Logout Whatsapp Account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Whatsapp"
                ],
                "summary": "Logout Whatsapp Account",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.MessageResponseSuccess"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.MessageResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.MessageResponseError"
                        }
                    }
                }
            }
        },
        "/wa/messages": {
            "post": {
                "description": "Send messages custom to whatsapp",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "News"
                ],
                "summary": "Send messages custom to whatsapp",
                "parameters": [
                    {
                        "description": "body request detail",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.WhatsappMessagesReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.MessageResponseSuccess"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.MessageResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.MessageResponseError"
                        }
                    }
                }
            }
        },
        "/wa/news": {
            "post": {
                "description": "Send news to whatsapp",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "News"
                ],
                "summary": "Send news to whatsapp",
                "parameters": [
                    {
                        "description": "body request detail",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.NewsSendWhatsappReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.MessageResponseSuccess"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.MessageResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.MessageResponseError"
                        }
                    }
                }
            }
        },
        "/wa/status": {
            "get": {
                "description": "Check Whatsapp Status",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Whatsapp"
                ],
                "summary": "Check Whatsapp Status",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.WAStatusResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.MessageResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.MessageResponseError"
                        }
                    }
                }
            }
        },
        "/wa/weathers": {
            "post": {
                "description": "Send weather daily forecast",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "News"
                ],
                "summary": "Send weather daily forecast to whatsapp",
                "parameters": [
                    {
                        "description": "body request detail",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.WeatherSendWhatsappReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.MessageResponseSuccess"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.MessageResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.MessageResponseError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handlers.WAStatusResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "properties": {
                        "is_connected": {
                            "type": "boolean"
                        },
                        "is_ready": {
                            "type": "boolean"
                        },
                        "qr_code": {
                            "type": "string"
                        }
                    }
                },
                "error": {
                    "type": "boolean",
                    "example": false
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "models.NewsSendWhatsappReq": {
            "type": "object",
            "properties": {
                "category": {
                    "description": "options: business, entertainment, general, health, science, sports, technology",
                    "type": "string",
                    "example": "business"
                },
                "whatsapp_numbers": {
                    "description": "list of numbers to send the news to and start with code number like 62 and not 0 like 08123456789",
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "6285727771234",
                        "6285667889887"
                    ]
                }
            }
        },
        "models.WeatherSendWhatsappReq": {
            "type": "object",
            "properties": {
                "lat": {
                    "description": "required, latitude of the location",
                    "type": "number",
                    "example": -6.2617
                },
                "lon": {
                    "description": "required, longitude of the location",
                    "type": "number",
                    "example": 106.8103
                },
                "type": {
                    "description": "required, options: today, tomorrow",
                    "type": "string",
                    "example": "today"
                },
                "whatsapp_numbers": {
                    "description": "list of numbers to send the news to and start with code number like 62 and not 0 like 08123456789",
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "6285727771234",
                        "6285667889887"
                    ]
                }
            }
        },
        "models.WhatsappMessagesReq": {
            "type": "object",
            "properties": {
                "messages": {
                    "description": "message to be sent to the whatsapp numbers",
                    "type": "string",
                    "example": "Hello, this is a test message"
                },
                "whatsapp_numbers": {
                    "description": "list of numbers to send the news to and start with code number like 62 and not 0 like 08123456789",
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "6285727771234",
                        "6285667889887"
                    ]
                }
            }
        },
        "utils.MessageResponseError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "boolean",
                    "default": true,
                    "example": true
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "utils.MessageResponseSuccess": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "boolean",
                    "default": false,
                    "example": false
                },
                "message": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    },
    "externalDocs": {
        "description": "OpenAPI",
        "url": "https://swagger.io/resources/open-api/"
    }
}